Refactored the markdown visual/block editor into it's own component
