+++ src/components/Widgets/MarkdownControlElements/constants.js
@@ -1,0 +1,13 @@
+export const emptyParagraphBlock = {
+  nodes: [
+    { kind: 'block',
+      type: 'paragraph',
+      nodes: [{
+        kind: 'text',
+        ranges: [{
+          text: ''
+        }]
+      }]
+    }
+  ]
+};
